pacsift: only strip leading root path with --exact
Always stripping the root path prevents searching for packages with
files containing it.  In particular, for installations with root '/',
stripping it prevents the use of '/' to match a full component name
(e.g. searching for '/bin' to match '/usr/bin' but not 'perlbin.sh').
For --exact searches it is safe to strip a leading root path because the
root path is always absolute and package paths are always relative.

Signed-off-by: Andrew Gregory <andrew.gregory.8@gmail.com>