pacrepairfile: add fchmodat wrapper
Linux does not provide a proper fchmodat call with AT_SYMLINK_NOFOLLOW
support, so glibc bails if that flag is used even if the path is not
a symlink.  Provide a wrapper function that tries the native fchmodat
call, but falls back to manually opening and manipulating a file
descriptor on that particular failure.

Signed-off-by: Andrew Gregory <andrew.gregory.8@gmail.com>