Check if stdin is a tty before initializing libalpm
Using the recommended pattern of invoking pacsift et al with <&-
causes the file descriptor 0 to not be valid. However, this also makes
it available for being allocated by calls to open or socket.

While this pattern worked as expected with previous versions of
pacman/libalpm, this is no longer true, as alpm_initialize now creates
a socket, which is allocated at file descriptor 0 if it is free. This
causes the isatty check to now fail with ENOTTY instead of EBADF,
causing the tools to think they were invoked with a list of packages
on their standard input.

Fix this by moving the stdin check at the top of each program, before
any other initialization. (For consistency, this was done whether or
not the program's initialization involves libalpm.)